refactor(secrets): split placeholder tests out, merge redundant case
placeholder.rs's #[cfg(test)] mod was ~745 lines, pushing the file to
1436 lines and past the repo's 1000-line convention despite the
production body itself being small. Move it to a sibling
placeholder_tests.rs via #[path], zero behavior change: super::
references still resolve since the module tree is identical, only the
file that backs it moves.

Also merge lease_revokes_on_explicit_success_call and
lease_revokes_on_explicit_error_call into one
lease_revokes_on_explicit_call: both called the identical
CredentialSessionLease::revoke API and asserted the identical
postcondition, differing only in a narrative
`if dispatch_result.is_err()` wrapper that touches no broker code
path. The timeout (cancellation-drop) and panic (unwind-drop) tests
stay separate since those exercise genuinely different Rust
mechanisms.

Add the missing lookup-coverage case: a stale (expired or
use-exhausted) session bound to a placeholder under the same scope as
a later, still-valid one must not prevent the valid one from being
found by find_session_by_placeholder. The existing multi-session test
only covered scope-mismatch skipping.
